{"types":"C","runs":53}

+++ problem16.java
@@ -6,7 +6,7 @@ public class problem16 extends ConsoleProgram {
 		double firstNum = (int) (readInt("First number is: "));
 		double secondNum = (int) (readInt("Second number is: "));
 		double thirdNum = (int) (readInt("Third number is : "));
-		int average = (int) (1 / (1 / firstNum + 1 / secondNum + 1 / thirdNum));
+		double average = 1 / (1 / firstNum + 1 / secondNum + 1 / thirdNum);
 		println("Average is: " + average);		
 	}
 }
